ARM: march-vt8500: create separate header file
This patch creates a separate header file for vt8500,
and move all the defines there. Related to this change,
the name of two defines were changed by adding a VT8500_
prefix to them to match the rest of the pack, and avoid
any unintended clash in the rest of the code.

Signed-off-by: Gergely Imreh <imrehg@gmail.com>